Add support for user resources and clarify context
I'm not sure this is the right direction to go with the project. I don't
exactly like how the user interface is turning out. For example, I don't
want the project's users to worry about locks and mutexes. The state
that they need should never be referenced by the firmware I write, so we
only need to worry about user-defined interrupts. Is there a way to make
the use of state variables more accessible?